update(driver-loader): Change dynlinker path
The Flatcar kernel module build needs to relocate binaries in /host/lib/modules to run with the
host's ld and glibc. While testing with glibc 2.34 that we plan to introduce soon, we found that
the previous dynlinker symlink (ld-xxx.so) doesn't exist any longer. Update the hostld variable to
use a path that is present in all versions of glibc.